requirements.txt
alabaster==0.7.12
alembic==1.4.0
amqp==2.5.2
aniso8601==8.0.0
apache-airflow==1.10.9
apispec==1.3.3
appdirs==1.4.3
argcomplete==1.11.1
aspy.yaml==1.3.0
astroid==2.3.3
attrs==19.3.0
aws-sam-translator==1.20.1
aws-xray-sdk==2.4.3
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
billiard==3.6.2.0
black==19.10b0
boto==2.49.0
boto3==1.11.16
botocore==1.14.16
cached-property==1.5.1
cachetools==4.0.0
cattrs==0.9.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
cfgv==3.0.0
cfn-lint==0.27.5
cgroupspy==0.1.6
chardet==3.0.4
click==6.7
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
coverage==5.0.3
croniter==0.3.31
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
dnspython==1.16.0
docker==3.7.3
docker-pycreds==0.4.0
docutils==0.16
ecdsa==0.15
entrypoints==0.3
eventlet==0.25.1
filelock==3.0.12
flake8==3.7.9
flake8-colors==0.1.6
flaky==3.6.1
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.0
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-OAuthlib==0.9.5
Flask-OpenID==1.2.5
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
flower==0.9.3
freezegun==0.3.14
funcsigs==1.0.2
future==0.16.0
gevent==1.4.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-bigtable==0.33.0
google-cloud-container==0.4.0
google-cloud-core==1.3.0
google-cloud-dlp==0.13.0
google-cloud-language==1.3.0
google-cloud-spanner==1.9.0
google-cloud-speech==1.3.2
google-cloud-storage==1.26.0
google-cloud-texttospeech==0.5.0
google-cloud-translate==2.0.1
google-cloud-videointelligence==1.12.1
google-cloud-vision==0.42.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphviz==0.13.2
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.27.1
grpcio-gcp==0.2.2
gunicorn==19.10.0
httplib2==0.17.0
identify==1.4.11
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==1.0.2
inflection==0.3.1
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
JayDeBeApi==1.1.1
jedi==0.16.0
Jinja2==2.10.3
jira==2.0.0
jmespath==0.9.4
JPype1==0.7.1
json-merge-patch==0.2
jsondiff==1.1.2
jsonpatch==1.25
jsonpickle==1.2
jsonpointer==2.0
jsonschema==3.2.0
kombu==4.6.7
kubernetes==10.0.1
lazy-object-proxy==1.4.3
librabbitmq==2.0.0
lockfile==0.12.2
Mako==1.1.1
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.2
mccabe==0.6.1
mock==4.0.1
mongomock==3.19.0
monotonic==1.5
more-itertools==8.2.0
moto==1.3.14
mypy==0.720
mypy-extensions==0.4.3
mysqlclient==1.3.14
nodeenv==1.3.5
ntlm-auth==1.4.0
numpy==1.18.1
oauthlib==2.1.0
packaging==20.1
pandas==1.0.0
pandas-gbq==0.13.0
parameterized==0.7.1
paramiko==2.7.1
parso==0.6.1
pathspec==0.7.0
pbr==5.4.4
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
pika==0.13.0
pluggy==0.13.1
pre-commit==2.0.1
presto-python-client==0.7.0
prison==0.1.2
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.6.7
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
PyAMQP==0.0.8.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pydata-google-auth==0.3.0
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
pysftp==0.2.9
PySocks==1.7.1
pytest==5.3.5
pytest-cov==2.8.1
pytest-instafail==0.4.1.post0
python-daemon==2.1.2
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
python-jose==3.1.0
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
pywinrm==0.4.1
PyYAML==5.3
qds-sdk==1.13.2
redis==3.4.1
regex==2020.2.20
requests==2.22.0
requests-mock==1.7.0
requests-ntlm==1.1.0
requests-oauthlib==1.1.0
requests-toolbelt==0.9.1
responses==0.10.9
rsa==4.0
s3transfer==0.3.3
sentinels==1.0.0
setproctitle==1.1.10
six==1.14.0
slackclient==1.3.2
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==2.4.1
sphinx-argparse==0.2.5
sphinx-autoapi==1.0.0
sphinx-jinja==1.1.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-dotnetdomain==0.4
sphinxcontrib-golangdomain==0.2.0.dev0
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.13
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.1
sshpubkeys==3.1.0
sshtunnel==0.1.5
tabulate==0.8.6
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
toml==0.10.0
tornado==5.1.1
traitlets==4.3.3
tweepy==3.8.0
typed-ast==1.4.1
typing==3.7.4.1
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
xmltodict==0.12.0
yamllint==1.20.0
zipp==2.2.0
zope.deprecation==4.4.0